refactor(app): 페이지 SSR 전환 및 Client 분리로 성능·SEO 개선
