+++ src/app/_components/features/job/EmptyEducations.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import Image from 'next/image';
 import Link from 'next/link';
