+++ src/app/education-programs/page.tsx
@@ -1,403 +1,155 @@
-'use client';
-
-import EducationCard from '@/components/features/job/EducationCard';
-import EmptyEducations from '@/components/features/job/EmptyEducations';
-import SearchBar from '@/components/ui/SearchBar';
-import EducationTab from '@/components/ui/EducationTab';
-import EducationFilter from '@/components/ui/EducationFilter';
-import EducationCardSkeleton from '@/components/ui/EducationCardSkeleton';
-import Footer from '@/components/layout/Footer';
-import { useState, useEffect } from 'react';
-import { getUserData, getAccessToken } from '@/lib/auth';
-import {
-  getRecommendedEducations,
-  getAllEducationsAnonymous,
-  getHrdEducations,
-  getEducationBookmarks,
-} from '@/lib/api/jobApi';
+import { cookies } from 'next/headers';
+import EducationProgramsClient from './EducationProgramsClient';
 import { EducationSummary } from '@/types/job';
 
-export default function EducationPrograms() {
-  const [favorites, setFavorites] = useState<Set<string>>(new Set());
-  const [activeTab, setActiveTab] = useState<'custom' | 'all'>('custom');
-  const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
-  const [isAuthChecked, setIsAuthChecked] = useState<boolean>(false);
-  const [educations, setEducations] = useState<EducationSummary[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [currentPage, setCurrentPage] = useState<number>(1);
-  const [totalPages, setTotalPages] = useState<number>(1);
-  const [totalElements, setTotalElements] = useState<number>(0);
-  const [searchKeyword, setSearchKeyword] = useState<string>('');
-  const [debouncedSearchKeyword, setDebouncedSearchKeyword] =
-    useState<string>('');
-  const [filters, setFilters] = useState<{
-    selectedRegion: string;
-    selectedDistricts: string[];
-    educationType: string[];
-    educationCategory: string[];
-  }>({
-    selectedRegion: '',
-    selectedDistricts: [],
-    educationType: [],
-    educationCategory: [],
-  });
+async function fetchInitialEducations(
+  isLoggedIn: boolean
+): Promise<{ educations: EducationSummary[]; totalElements: number }> {
+  const backendUrl = process.env.NEXT_PUBLIC_BACKEND_URL;
+  if (!backendUrl) {
+    return { educations: [], totalElements: 0 };
+  }
 
-  // 초기 로그인 상태 확인 및 데이터 로드
-  useEffect(() => {
-    const userData = getUserData();
-    const token = getAccessToken();
-    const loggedIn = !!(userData?.name && token);
-
-    setIsLoggedIn(loggedIn);
-    setIsAuthChecked(true);
-
-    // 비로그인 시 전체교육 탭으로 설정, 로그인 시 맞춤교육으로 설정
-    if (!loggedIn) {
-      setActiveTab('all');
-    } else {
-      setActiveTab('custom');
-    }
-
-    // 로그인된 경우 찜 목록 로드
-    if (loggedIn) {
-      loadBookmarks();
-    }
-  }, []);
-
-  // 찜 목록 로드 함수
-  const loadBookmarks = async () => {
-    try {
-      const bookmarkIds = await getEducationBookmarks();
-      setFavorites(new Set(bookmarkIds));
-    } catch (error) {
-      console.error('Error loading education bookmarks:', error);
-    }
-  };
-
-  // 검색어 디바운싱 (500ms 지연)
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebouncedSearchKeyword(searchKeyword);
-    }, 500);
-
-    return () => clearTimeout(timer);
-  }, [searchKeyword]);
-
-  // 데이터 로드 함수
-  const fetchEducations = async (page: number = 1) => {
-    try {
-      setIsLoading(true);
-      let educationData: EducationSummary[] = [];
-      let totalElements = 0;
-
-      if (isLoggedIn) {
-        // 로그인 시
-        if (activeTab === 'custom') {
-          // 맞춤 교육: /education/recommend API 사용
-          educationData = await getRecommendedEducations();
-          totalElements = educationData.length; // 맞춤 교육은 페이지네이션이 없을 수 있음
-        } else {
-          // 전체 교육: /education/hrd-course API 사용
-          const result = await getHrdEducations({
-            keyword: debouncedSearchKeyword || undefined,
-            pageNo: page,
-            pageSize: 10,
-            startYmd: '20250101',
-            endYmd: '20251231',
-          });
-          // EducationDto를 EducationSummary 형식으로 변환
-          educationData = (result.educationDtoList || []).map((edu) => ({
-            id: edu.educationId.toString(),
-            educationId: edu.educationId,
-            trprId: edu.educationId.toString(),
-            title: edu.title,
-            subTitle: edu.subTitle,
-            institution: '',
-            address: edu.address,
-            traStartDate: edu.traStartDate,
-            traEndDate: edu.traEndDate,
-            trainTarget: '',
-            contents: '',
-            certificate: '',
-            grade: '',
-            regCourseMan: '',
-            courseMan: edu.courseMan,
-            realMan: '',
-            yardMan: '',
-            telNo: '',
-            stdgScor: '',
-            eiEmplCnt3: '',
-            eiEmplRate3: '',
-            eiEmplCnt3Gt10: '',
-            eiEmplRate6: '',
-            ncsCd: '',
-            trprDegr: edu.trprDegr,
-            instCd: '',
-            trngAreaCd: '',
-            trainTargetCd: '',
-            trainstCstId: '',
-            subTitleLink: '',
-            titleLink: edu.titleLink,
-            titleIcon: '',
-            imageUrl: edu.imageUrl,
-            isBookmark: edu.isBookmark,
-            recommendScore: edu.score ?? undefined,
-          }));
-          totalElements = result.totalElements || 0;
-        }
-      } else {
-        // 비로그인 시: /education/all/anonymous API 사용
-        const result = await getAllEducationsAnonymous({
-          keyword: debouncedSearchKeyword || undefined,
-          startYmd: '',
-          endYmd: '',
-        });
-        // EducationDto를 EducationSummary 형식으로 변환
-        educationData = (result.educationDtoList || []).map((edu) => ({
-          id: edu.educationId.toString(),
-          educationId: edu.educationId,
-          trprId: edu.educationId.toString(),
-          title: edu.title,
-          subTitle: edu.subTitle,
-          institution: '',
-          address: edu.address,
-          traStartDate: edu.traStartDate,
-          traEndDate: edu.traEndDate,
+  try {
+    if (isLoggedIn) {
+      const accessToken = (await cookies()).get('accessToken')?.value;
+      if (!accessToken) {
+        return { educations: [], totalElements: 0 };
+      }
+      // 초기 탭은 로그인 시 custom(추천)
+      const response = await fetch(`${backendUrl}/education/recommend`, {
+        method: 'GET',
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+          'Content-Type': 'application/json',
+        },
+        cache: 'no-store',
+      });
+      if (!response.ok) {
+        return { educations: [], totalElements: 0 };
+      }
+      const result = await response.json();
+      if (result.result !== 'SUCCESS') {
+        return { educations: [], totalElements: 0 };
+      }
+      const educations: EducationSummary[] = (result.data.jobDtoList || []).map(
+        (item: any) => ({
+          id: String(item.jobId ?? ''),
+          educationId: Number(item.jobId ?? 0),
+          trprId: String(item.jobId ?? ''),
+          title: item.jobTitle || '제목 없음',
+          subTitle: item.jobCodeName || '',
+          institution: item.companyName || '',
+          address: item.workLocation || '',
+          traStartDate: item.postingDate || '',
+          traEndDate: item.closingDate || '',
           trainTarget: '',
-          contents: '',
+          contents: item.description || '',
           certificate: '',
           grade: '',
-          regCourseMan: '',
-          courseMan: edu.courseMan,
-          realMan: '',
-          yardMan: '',
-          telNo: '',
-          stdgScor: '',
-          eiEmplCnt3: '',
-          eiEmplRate3: '',
-          eiEmplCnt3Gt10: '',
-          eiEmplRate6: '',
+          regCourseMan: String(item.recruitNumber ?? '0'),
+          courseMan: String(item.recruitNumber ?? '0'),
+          realMan: '0',
+          yardMan: '0',
+          telNo: item.managerPhone || '',
+          stdgScor: '0',
+          eiEmplCnt3: '0',
+          eiEmplRate3: '0',
+          eiEmplCnt3Gt10: '0',
+          eiEmplRate6: '0',
           ncsCd: '',
-          trprDegr: edu.trprDegr,
+          trprDegr: '',
           instCd: '',
           trngAreaCd: '',
           trainTargetCd: '',
           trainstCstId: '',
           subTitleLink: '',
-          titleLink: edu.titleLink,
+          titleLink: '',
           titleIcon: '',
-          imageUrl: edu.imageUrl,
-          isBookmark: edu.isBookmark,
-          recommendScore: edu.score ?? undefined,
-        }));
-        totalElements = result.totalElements || 0;
-      }
-
-      setEducations(educationData);
-      setTotalElements(totalElements);
-      setTotalPages(Math.ceil(totalElements / 10));
-    } catch (error) {
-      console.error('Error fetching educations:', error);
-      setEducations([]);
-      setTotalElements(0);
-      setTotalPages(1);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // 탭, 필터, 검색어 변경 시 첫 페이지로 데이터 로드
-  useEffect(() => {
-    if (!isAuthChecked) return;
-
-    setCurrentPage(1);
-    fetchEducations(1);
-  }, [activeTab, isLoggedIn, filters, debouncedSearchKeyword, isAuthChecked]);
-
-  // 페이지 변경 시 데이터 로드
-  useEffect(() => {
-    if (currentPage > 1 && isAuthChecked) {
-      fetchEducations(currentPage);
-    }
-  }, [currentPage]);
-
-  const toggleFavorite = (educationId: string) => {
-    const newFavorites = new Set(favorites);
-    if (newFavorites.has(educationId)) {
-      newFavorites.delete(educationId);
+          imageUrl: item.imageUrl,
+          isBookmark: Boolean(item.isBookmark),
+          recommendScore: item.score,
+        })
+      );
+      return { educations, totalElements: educations.length };
     } else {
-      newFavorites.add(educationId);
+      // 비로그인 시: 내부 API 라우트를 통해 익명 전체 조회 (페이지 1, size 10)
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_BASE_PATH || ''}/api/education/anonymous?pageNo=1&pageSize=10`,
+        {
+          method: 'GET',
+          headers: { 'Content-Type': 'application/json' },
+          cache: 'no-store',
+        }
+      );
+      if (!response.ok) {
+        return { educations: [], totalElements: 0 };
+      }
+      const result = await response.json();
+      if (result.result !== 'SUCCESS') {
+        return { educations: [], totalElements: 0 };
+      }
+      const data = result.data;
+      const educations: EducationSummary[] = (data.educationDtoList || []).map(
+        (edu: any) => ({
+          id: String(edu.educationId),
+          educationId: Number(edu.educationId),
+          trprId: String(edu.educationId),
+          title: edu.title || '제목 없음',
+          subTitle: edu.subTitle || '',
+          institution: edu.subTitle || '',
+          address: edu.address || '',
+          traStartDate: edu.traStartDate || '',
+          traEndDate: edu.traEndDate || '',
+          trainTarget: '',
+          contents: edu.keyword1 || edu.keyword2 || '',
+          certificate: '',
+          grade: '',
+          regCourseMan: '0',
+          courseMan: edu.courseMan || '0',
+          realMan: '0',
+          yardMan: '0',
+          telNo: '',
+          stdgScor: '0',
+          eiEmplCnt3: '0',
+          eiEmplRate3: '0',
+          eiEmplCnt3Gt10: '0',
+          eiEmplRate6: '0',
+          ncsCd: '',
+          trprDegr: edu.trprDegr || '',
+          instCd: '',
+          trngAreaCd: '',
+          trainTargetCd: '',
+          trainstCstId: '',
+          subTitleLink: '',
+          titleLink: edu.titleLink || '',
+          titleIcon: '',
+          imageUrl: edu.imageUrl || '',
+          isBookmark: Boolean(edu.isBookmark),
+          recommendScore: edu.score || undefined,
+        })
+      );
+      return {
+        educations,
+        totalElements: data.totalElements || educations.length,
+      };
     }
-    setFavorites(newFavorites);
-  };
-
-  if (isLoading) {
-    return (
-      <div>
-        <main className="min-h-screen bg-white">
-          <section className="w-full px-4 py-8">
-            <div className="max-w-[1200px] mx-auto">
-              <SearchBar />
-
-              <EducationFilter onFilterChange={setFilters} />
-
-              {isLoggedIn && (
-                <EducationTab
-                  activeTab={activeTab}
-                  onTabChange={setActiveTab}
-                  isLoggedIn={isLoggedIn}
-                />
-              )}
-
-              <div className="flex flex-col md:flex-row gap-6 mt-12">
-                <div className="flex flex-col gap-6 flex-1">
-                  {Array.from({ length: 4 }).map((_, index) => (
-                    <EducationCardSkeleton key={index} />
-                  ))}
-                </div>
-                <div className="flex flex-col gap-6 flex-1">
-                  {Array.from({ length: 4 }).map((_, index) => (
-                    <EducationCardSkeleton key={index + 4} />
-                  ))}
-                </div>
-              </div>
-            </div>
-          </section>
-        </main>
-        <Footer />
-      </div>
-    );
+  } catch (_e) {
+    return { educations: [], totalElements: 0 };
   }
+}
+
+export default async function EducationPrograms() {
+  const cookieStore = await cookies();
+  const isLoggedIn = Boolean(cookieStore.get('accessToken')?.value);
+  const { educations, totalElements } =
+    await fetchInitialEducations(isLoggedIn);
 
   return (
-    <div>
-      <main className="min-h-screen bg-white">
-        <section className="w-full px-4 py-8">
-          <div className="max-w-[1200px] mx-auto">
-            <SearchBar onSearchChange={setSearchKeyword} />
-
-            <EducationFilter onFilterChange={setFilters} />
-
-            {isLoggedIn && (
-              <EducationTab
-                activeTab={activeTab}
-                onTabChange={setActiveTab}
-                isLoggedIn={isLoggedIn}
-              />
-            )}
-
-            {/* 빈 상태 처리 - 로그인 상태에서만 표시 */}
-            {educations.length === 0 && isLoggedIn ? (
-              <EmptyEducations isLoggedIn={isLoggedIn} activeTab={activeTab} />
-            ) : educations.length > 0 ? (
-              <>
-                <div className="flex flex-col md:flex-row gap-6 mt-12">
-                  <div className="flex flex-col gap-6 flex-1">
-                    {educations
-                      .slice(0, Math.ceil(educations.length / 2))
-                      .map((education, index) => (
-                        <EducationCard
-                          key={education.trprId || index}
-                          education={education}
-                          onToggleBookmark={toggleFavorite}
-                          isBookmarked={favorites.has(education.trprId || '')}
-                        />
-                      ))}
-                  </div>
-                  <div className="flex flex-col gap-6 flex-1">
-                    {educations
-                      .slice(Math.ceil(educations.length / 2))
-                      .map((education, index) => (
-                        <EducationCard
-                          key={
-                            education.trprId ||
-                            index + Math.ceil(educations.length / 2)
-                          }
-                          education={education}
-                          onToggleBookmark={toggleFavorite}
-                          isBookmarked={favorites.has(education.trprId || '')}
-                        />
-                      ))}
-                  </div>
-                </div>
-                {/* 페이지네이션 */}
-                {totalPages > 1 && (
-                  <div className="flex justify-center items-center gap-2 mt-12 mb-8">
-                    {/* 첫 페이지 버튼 */}
-                    <button
-                      onClick={() => setCurrentPage(1)}
-                      disabled={currentPage === 1}
-                      className="px-3 py-2 rounded-lg border border-gray-300 text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      &lt;&lt;
-                    </button>
-
-                    {/* 이전 버튼 */}
-                    <button
-                      onClick={() =>
-                        setCurrentPage(Math.max(1, currentPage - 1))
-                      }
-                      disabled={currentPage === 1}
-                      className="px-4 py-2 rounded-lg border border-gray-300 text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      &lt;
-                    </button>
-
-                    {/* 페이지 번호들 */}
-                    {Array.from({ length: Math.min(5, totalPages) }, (_, i) => {
-                      let pageNum;
-                      if (totalPages <= 5) {
-                        pageNum = i + 1;
-                      } else if (currentPage <= 3) {
-                        pageNum = i + 1;
-                      } else if (currentPage >= totalPages - 2) {
-                        pageNum = totalPages - 4 + i;
-                      } else {
-                        pageNum = currentPage - 2 + i;
-                      }
-
-                      return (
-                        <button
-                          key={pageNum}
-                          onClick={() => setCurrentPage(pageNum)}
-                          className={`min-w-[40px] px-4 py-2 rounded-lg border font-medium ${
-                            currentPage === pageNum
-                              ? 'bg-primary-90 text-white border-primary-90 shadow-md'
-                              : 'border-gray-300 text-gray-700 hover:bg-gray-50 hover:border-gray-400'
-                          }`}
-                        >
-                          {pageNum}
-                        </button>
-                      );
-                    })}
-
-                    {/* 다음 버튼 */}
-                    <button
-                      onClick={() =>
-                        setCurrentPage(Math.min(totalPages, currentPage + 1))
-                      }
-                      disabled={currentPage === totalPages}
-                      className="px-4 py-2 rounded-lg border border-gray-300 text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      &gt;
-                    </button>
-
-                    {/* 마지막 페이지 버튼 */}
-                    <button
-                      onClick={() => setCurrentPage(totalPages)}
-                      disabled={currentPage === totalPages}
-                      className="px-3 py-2 rounded-lg border border-gray-300 text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      &gt;&gt;
-                    </button>
-                  </div>
-                )}
-              </>
-            ) : null}
-          </div>
-        </section>
-      </main>
-      <Footer />
-    </div>
+    <EducationProgramsClient
+      initialEducations={educations}
+      initialTotalElements={totalElements}
+      isLoggedInInitial={isLoggedIn}
+    />
   );
 }
